Remove vc-auth in favor of a single validator file.
  - Replace security context with credentials context in peerDeps.
