+++ lib/http.js
@@ -33,7 +33,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
   const {routes} = cfg;
 
   app.post(routes.authenticate,
-    validate({body: 'vc-auth.login'}),
+    validate({body: 'vc-issuer.login'}),
     asyncHandler(async (req, res) => {
       const {body: {presentation}} = req;
       // extract the DID from the presentation
@@ -133,7 +133,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
       const {controller, presentation: vp} = req.body;
 
       // disallow any authenticated user to act as another controller
-      const {actor} = req.user || {};
+      const {actor} = req.user;
       if(actor.id !== controller) {
         throw new BedrockError(
           'Permission denied.',
@@ -208,7 +208,7 @@ bedrock.events.on('bedrock-express.configure.routes', app => {
   app.post(
     routes.claimUser,
     ensureAuthenticated,
-    validate({body: 'vc-auth.claimUser'}),
+    validate({body: 'vc-issuer.claimUser'}),
     asyncHandler(async (req, res) => {
       const {body, user: {account}} = req;
       const {instanceId, token} = body;
