+++ test/package.json
@@ -16,7 +16,7 @@
     "bedrock-mongodb": "^6.0.2",
     "bedrock-passport": "^5.0.2",
     "bedrock-permission": "^2.5.4",
-    "bedrock-security-context": "^3.0.0",
+    "bedrock-credentials-context": "^1.0.0",
     "bedrock-server": "^2.4.1",
     "bedrock-test": "^5.0.0",
     "bedrock-validation": "^4.2.0",
